Add checks

+++ lenovo-100w-gen3/install.sh
@@ -1,6 +1,9 @@
 installconfig () {
     sudo pacman -Syu hyprlock hypridle fontforge python
     echo "Linking hypr files into config directory..."
+
+    rm ~/.config/hypr/hyprland.conf
+
     readarray -d '' array < <(find . -name hypr*.conf -print0)
 
     for file in "${array[@]}"; do
@@ -51,6 +54,9 @@ installevalockscreen () {
     sh ./usr/local/share/fonts/ttf/EVA-Matisse/fontextracter.sh
 
     echo "Linking hyprlock conf..."
+
+    rm ~/.config/hypr/hyprlock.conf
+
     filelink="$(pwd | tr -d '\n')"
     ln -s $filelink/hypr/hyprlock.conf $HOME/.config/hypr/hyprlock.conf
 
@@ -61,38 +67,42 @@ installevalockscreen () {
 }
 
 fixtrackpad () {
-    sudo pacman -Syu acpica cpio
-    cd /tmp
+    if grep 'GRUB_EARLY_INITRD_LINUX_CUSTOM="acpi_override"' /etc/default/grub; then
+        echo "Fix already installed!"
+    elif ! grep 'GRUB_EARLY_INITRD_LINUX_CUSTOM="acpi_override"' /etc/default/grub; then
+        sudo pacman -Syu acpica cpio
+        cd /tmp
 
-    echo "Grabbing and decompiling DSDT file..."
-    sudo cat /sys/firmware/acpi/tables/DSDT > dsdt.aml
-    iasl -d dsdt.aml
+        echo "Grabbing and decompiling DSDT file..."
+        sudo cat /sys/firmware/acpi/tables/DSDT > dsdt.aml
+        iasl -d dsdt.aml
 
-    echo "Replacing text..."
-    sed -i -z 's/If ((^^^PCI0.LPC0.H_EC.ECRD (RefOf (^^^PCI0.LPC0.H_EC.TPTY)) == 0x02))/Else/2' dsdt.dsl
-    sed -i -z 's:CDAT://CDAT:8' dsdt.dsl
-    echo "Compiling DSDT file..."
-    iasl -sa dsdt.dsl
+        echo "Replacing text..."
+        sed -i -z 's/If ((^^^PCI0.LPC0.H_EC.ECRD (RefOf (^^^PCI0.LPC0.H_EC.TPTY)) == 0x02))/Else/2' dsdt.dsl
+        sed -i -z 's:CDAT://CDAT:8' dsdt.dsl
+        echo "Compiling DSDT file..."
+        iasl -sa dsdt.dsl
 
-    echo "Setting up CPIO table..."
-    mkdir -p ~/kernel/firmware/acpi
-    cd ~ && cp /tmp/dsdt.aml ~/kernel/firmware/acpi
-    find kernel | cpio -H newc --create > acpi_override
-    sudo cp acpi_override /boot
+        echo "Setting up CPIO table..."
+        mkdir -p ~/kernel/firmware/acpi
+        cd ~ && cp /tmp/dsdt.aml ~/kernel/firmware/acpi
+        find kernel | cpio -H newc --create > acpi_override
+        sudo cp acpi_override /boot
 
-    echo "Configurating GRUB..."
-    sudo sed -i '8iGRUB_EARLY_INITRD_LINUX_CUSTOM="acpi_override"' /etc/default/grub
-    sudo grub-mkconfig -o /boot/grub/grub.cfg
+        echo "Configurating GRUB..."
+        sudo sed -i '8iGRUB_EARLY_INITRD_LINUX_CUSTOM="acpi_override"' /etc/default/grub
+        sudo grub-mkconfig -o /boot/grub/grub.cfg
 
-    echo -e "\nYou need to restart for the trackpad fix to apply. Would you like to do that now? (Y/n)"
-    read answer
+        echo -e "\nYou need to restart for the trackpad fix to apply. Would you like to do that now? (Y/n)"
+        read answer
 
-    if [[ "${answer,,}" == 'y' ]]; then
-        sudo reboot
-    elif [[ "${answer,,}" == 'n' ]]; then
-        exit
-    elif [[ "${answer,,}" == '' ]]; then
-        sudo reboot
+        if [[ "${answer,,}" == 'y' ]]; then
+            sudo reboot
+        elif [[ "${answer,,}" == 'n' ]]; then
+            exit
+        elif [[ "${answer,,}" == '' ]]; then
+            sudo reboot
+        fi
     fi
 }
 
@@ -122,21 +132,67 @@ done
 for i in "${answerarray[@]}"; do
     case $i in
         "1")
-            installconfig
+            if [ -L "${HOME}/.config/hypr/hyprland.conf" ]; then
+                echo "Links already exists!"
+            elif [ -f "${HOME}/.config/hypr/hyprland.conf" ] || [ ! -f "${HOME}/.config/hypr/hyprland.conf"]; then 
+                echo "This will install/overwrite your current hypr confs. Are you sure you want to install this? (y/N)"
+                read answer
+                if [[ "${answer,,}" == 'y' ]]; then
+                    installconfig
+                elif [[ "${answer,,}" == 'n' ]] || [[ "${answer,,}" == '' ]]; then
+                    echo "Skipping config install..."
+                fi
+            fi
         ;;
         "2")
-            installpowersaver
+            if [ -f "/etc/udev/rules.d/99-powersave.rules" ]; then
+                echo "Powersave already exists!"
+            elif [ ! -f "/etc/udev/rules.d/99-powersave.rules" ]; then 
+                echo "This will install a powersave feature. Are you sure you want to install this? (y/N)"
+                read answer
+                if [[ "${answer,,}" == 'y' ]]; then
+                    installpowersaver
+                elif [[ "${answer,,}" == 'n' ]] || [[ "${answer,,}" == '' ]]; then
+                    echo "Skipping powersave install..."
+                fi
+            fi
         ;;
         "3")
-            installevalockscreen
+            if [ -L "${HOME}/.config/hypr/hyprlock.conf" ]; then
+                echo "Lockscreen conf already exists!"
+            elif [ -f "${HOME}/.config/hypr/hyprlock.conf" ]; then 
+                echo "This will install the EVA lockscreen. Are you sure you want to install this? (y/N)"
+                read answer
+                if [[ "${answer,,}" == 'y' ]]; then
+                    installevalockscreen
+                elif [[ "${answer,,}" == 'n' ]] || [[ "${answer,,}" == '' ]]; then
+                    echo "Skipping powersave install..."
+                fi
+            fi
         ;;
         "4")
-            fixtrackpad
+            if grep 'GRUB_EARLY_INITRD_LINUX_CUSTOM="acpi_override"' /etc/default/grub; then
+                echo "Fix already installed!"
+            elif ! grep 'GRUB_EARLY_INITRD_LINUX_CUSTOM="acpi_override"' /etc/default/grub; then
+                echo "This will install the trackpad fix. Are you sure you want to install it? (y/N)"
+                read answer
+                if [[ "${answer,,}" == 'y' ]]; then
+                    fixtrackpad
+                elif [[ "${answer,,}" == 'n' ]] || [[ "${answer,,}" == '' ]]; then
+                    echo "Skipping trackpad fix..."
+                fi
+            fi
         ;;
         "A")
-            installconfig
-            installpowersaver
-            fixtrackpad
+            echo "This will install EVERYTHING. Are you sure? (y/N)"
+            read answer
+            if [[ "${answer,,}" == 'y' ]]; then
+                installconfig
+                installpowersaver
+                fixtrackpad
+            elif [[ "${answer,,}" == 'n' ]] || [[ "${answer,,}" == '' ]]; then
+                echo "Skipping install of everything..."
+            fi
         ;;
         "E")
             exit
